+++ src/audio/audio_capture.h
@@ -3,9 +3,9 @@
 #include <string>
 #include <vector>
 #include <memory>
-#include "audio_format.h"
+#include "common/model_config.h"
 
-namespace voice {
+namespace audio {
 
 class IAudioCapture {
 public:
@@ -15,7 +15,7 @@ public:
     virtual bool initialize() = 0;
 
     // 开始录制指定应用的音频
-    virtual bool start_recording_application(uint32_t app_id, const std::string& output_path = "") = 0;
+    virtual bool start_recording_application(uint32_t app_id) = 0;
 
     // 停止录制
     virtual void stop_recording() = 0;
@@ -24,7 +24,7 @@ public:
     virtual void list_applications() = 0;
 
     // 工厂方法创建平台特定实现
-    static std::unique_ptr<IAudioCapture> create(const std::string& config_path = "");
+    static std::unique_ptr<IAudioCapture> create(const common::ModelConfig& config);
 };
 
 } // namespace voice
