+++ src/translator/deepl/deeplx_translator.h
@@ -3,10 +3,12 @@
 #include <string>
 #include <memory>
 #include <system_error>
+#include "translator/deepl/deeplx_translator.h"
+#include "translator/translator.h"
 
-namespace voice {
+namespace deeplx {
 
-class DeepLXTranslator {
+class DeepLXTranslator : public translator::ITranslator {
 public:
     struct Config {
         std::string url;
@@ -14,11 +16,11 @@ public:
         std::string target_lang;
     };
 
-    explicit DeepLXTranslator(const Config& config);
-    ~DeepLXTranslator();
+    explicit DeepLXTranslator(const common::ModelConfig& config);
+    ~DeepLXTranslator() override = default;
 
     // Translate text if source language is different from target language
-    std::string translate(const std::string& text, const std::string& source_lang);
+    std::string translate(const std::string& text, const std::string& source_lang) override;
 
 private:
     struct HttpResponse {
@@ -37,4 +39,4 @@ private:
     int port_;
 };
 
-} // namespace voice 
+} // namespace deeplx
